The registery.parflow was added:
It seems that I forgot to add the registery.parflow file to the registery folder. The compilation problem might be due to this. As the grep of the PF_SATURATION brought back the registery.parflow as one of the files to have the phrase. However, I did not have that in my Registery folder.