The log.compile file of the successful compilation is addaed:
Finally the code was successfully compiled. All the executable files are created in the Main folder. It was compiled on a local machine (Ubuntu 16.04) with gfortran. The only Error that still exist is the error with regards to diffwrf which has been around in all my compilation. It seems to be not in the way of generating the executables so I always ignored it.